requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.1
python-digitalocean==1.17.0
loguru==0.7.2
aiohttp==3.9.1
openai==1.7.2
pydantic==2.7.2
numpy==1.26.4
pandas==2.2.0
chromadb==0.4.24
duckduckgo-search==4.1.1
jsonschema==4.21.1
python-telegram-bot==20.7
